kymotracker: split kymotracker test file
