+++ lumicks/pylake/kymotracker/tests/test_kymoline.py
@@ -260,3 +260,14 @@ def test_binding_histograms():
     counts, edges = lines._histogram_binding_events("all", bins=[2, 3, 4, 5, 6, 7, 8])
     np.testing.assert_equal(counts, [1, 2, 3, 3, 2, 1])
     np.testing.assert_allclose(edges, [2, 3, 4, 5, 6, 7, 8])
+
+
+def test_kymoline_regression_sample_from_image_clamp():
+    """This tests for a regression that occurred in sample_from_image. When sampling the image, we
+    sample pixels in a region around the line. This sampling procedure is constrained to stay within
+    the image. Previously, we used the incorrect axis to clamp the coordinate.
+    """
+    # Sampling the bottom row of a three pixel tall image will return [0, 0] instead of [1, 3];
+    # since both coordinates would be clamped to the edge of the image (sampling nothing)."""
+    img = CalibratedKymographChannel("test_data", np.array([[1, 1, 1], [3, 3, 3]]).T, 1e9, 1)
+    assert np.array_equal(KymoLine([0, 1], [2, 2], img).sample_from_image(0), [1, 3])
